// apps/web/src/proto/Challenge.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ServerInfo } from "./ServerInfo";

export type Challenge = { protocol: number, 
/**
 * Base64url random nonce to sign.
 */
nonce: string, server: ServerInfo, password_required: boolean, };